requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
asttokens==3.0.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
billiard==4.2.2
blis==1.3.0
build==1.3.0
cachetools==6.2.0
catalogue==2.0.10
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
chromadb==1.1.1
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.22.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
concurrent-log-handler==0.9.29
confection==0.1.5
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==46.0.2
cymem==2.0.11
daphne==4.2.1
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
Django==5.2.6
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-pytest==0.2.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
dnspython==2.8.0
durationpy==0.10
emoji==2.15.0
executing==2.2.1
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.9.0
google-auth==2.41.1
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==24.7.2
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kombu==5.5.4
kubernetes==34.1.0
langchain==1.2.14
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.2.23
langchain-ollama==1.0.0
langchain-openai==1.1.12
langchain-text-splitters==1.0.0
langcodes==3.5.0
langdetect==1.0.9
langgraph==1.1.4
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.4.33
language_data==1.3.0
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
murmurhash==1.0.13
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.3.3
oauthlib==3.3.1
olefile==0.47
ollama==0.6.0
onnxruntime==1.23.1
openai==2.30.0
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.7
ormsgpack==1.11.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
portalocker==3.2.0
posthog==5.4.0
preshed==3.0.10
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.1
psycopg2==2.9.11
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.15.3
pyOpenSSL==25.3.0
pypdf==6.1.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==9.0.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.1
redis==6.4.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
schedule==1.2.2
scikit-learn==1.6.0
scipy==1.16.2
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
Twisted==25.5.0
txaio==25.9.2
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.18.15
unstructured-client==0.42.3
urllib3==2.3.0
uuid==1.30
uuid_utils==0.14.1
uvicorn==0.37.0
vine==5.1.0
wasabi==1.1.3
watchfiles==1.1.0
wcwidth==0.2.14
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zope.interface==8.0.1
zstandard==0.25.0